Qt6 Build support (#898)
Qt6 is now supported while maintaining the ability to build with Qt5!

+++ plugins/HeapAnalyzer/DialogHeap.cpp
@@ -215,10 +215,10 @@ DialogHeap::DialogHeap(QWidget *parent, Qt::WindowFlags f)
 					}
 				}
 			}
-			qDebug("[Heap Analyzer] Done Processing %d Nodes", nodes.size());
+			qDebug("[Heap Analyzer] Done Processing %d Nodes", static_cast<int>(nodes.size()));
 
 			if (nodes.size() > MaxNodes) {
-				qDebug("[Heap Analyzer] Too Many Nodes! (%d)", nodes.size());
+				qDebug("[Heap Analyzer] Too Many Nodes! (%d)", static_cast<int>(nodes.size()));
 				delete graph;
 				return;
 			}
